Intro helpers & middleware
