+++ helpers/getFile.js
@@ -1,0 +1,6 @@
+
+const getFile = (originalFilename) => {
+  return originalFilename.split("|").pop();
+}
+
+module.exports = getFile;
